third_party/WebKit/LayoutTests/accessibility/br-element-has-correct-title.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>

<p>A br element<br id="br">should return newline as its accessible title.</p>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that a br element will return the newline character as its accessible title.");

    if (window.accessibilityController) {
        var brElement = accessibilityController.accessibleElementById("br");
        shouldBe("brElement.name", "'\\n'");
    }
</script>

</body>
</html>
